legacy/requirements.txt
pyzmq
gevent
pyserial
python-twitter
requests
Flask
influxdb
